Updates galore!
* Updated to Rdio iOS SDK 3.0.3
* Updated project target for iOS 8.3

Todo:
- [ ] Implement proper OAuth 2.0 libraries

+++ classes/Common.m
@@ -45,7 +45,7 @@
     }
     else {
         // First try and remove the extended attribute if it is present
-        int result = getxattr(filePath, attrName, NULL, sizeof(u_int8_t), 0, 0);
+        int result = (int)getxattr(filePath, attrName, NULL, sizeof(u_int8_t), 0, 0);
         if (result != -1) {
             // The attribute exists, we need to remove it
             int removeResult = removexattr(filePath, attrName, 0);
